Complete project
Unfortunately, I did not add any commit while creating this project. The
reason for this was that it was not mentioned in the odin project
instructions. At the end of the instructions, it mentioned that I would
learn how to deploy my app in the next lesson, so I assumed that
I was meant to skip adding any commits. Git is a version control system
that is meant to keep track of project progress. It is not separate from
app deployment. I learned my lesson, I won't make the same mistake in
the future.