Add favicon and remove required message
I installed the serve-favicon middleware to include a favicon in this
Express app. With regards to the required message, this is a pseudo
element I created in the new message form that appears when the user
clicks submit and did not fill the required fields. I realized before
deploying my app that, because the messages are absolutely positioned,
they do not appear in the appropriate positions for all screen sizes. I
decided to comment out this feature, and will potentially update this in
the future. The focus right now is to practice deploying an Express app
with a PostgreSQL database, so I did not want to waste any more time
working on this.